Add basic order types
