+++ .gitignore
@@ -2,6 +2,5 @@
 node_modules
 dist
 .env
-json-spec
 examples/*/yarn.lock
 examples/*/package-lock.json
